+++ axle-core/src/main/scala/axle/stats/ConditionalProbabilityTable.scala
@@ -14,24 +14,24 @@ import spire.random.Generator
 
 import axle.dummy
 
-object ConditionalProbabilityTable0 {
+object ConditionalProbabilityTable {
 
-  implicit def showCPT[A: Show: Order, V: Show: Field](implicit prob: ProbabilityModel[ConditionalProbabilityTable0]): Show[ConditionalProbabilityTable0[A, V]] = cpt =>
+  implicit def showCPT[A: Show: Order, V: Show: Field](implicit prob: ProbabilityModel[ConditionalProbabilityTable]): Show[ConditionalProbabilityTable[A, V]] = cpt =>
     cpt.values.sorted.map(a => {
       val aString = Show[A].show(a)
       (aString + (1 to (cpt.charWidth - aString.length)).map(i => " ").mkString("") + " " + Show[V].show(prob.probabilityOf(cpt, a)))
     }).mkString("\n")
 
-  implicit val probabilityWitness: ProbabilityModel[ConditionalProbabilityTable0] =
-    new ProbabilityModel[ConditionalProbabilityTable0] {
+  implicit val probabilityWitness: ProbabilityModel[ConditionalProbabilityTable] =
+    new ProbabilityModel[ConditionalProbabilityTable] {
 
-      def construct[A, V](variable: Variable[A], as: Iterable[A], f: A => V)(implicit ring: Ring[V]): ConditionalProbabilityTable0[A, V] =
-        ConditionalProbabilityTable0[A, V](as.map(a => a -> f(a)).toMap, variable)
+      def construct[A, V](variable: Variable[A], as: Iterable[A], f: A => V)(implicit ring: Ring[V]): ConditionalProbabilityTable[A, V] =
+        ConditionalProbabilityTable[A, V](as.map(a => a -> f(a)).toMap, variable)
 
-      def values[A](model: ConditionalProbabilityTable0[A, _]): IndexedSeq[A] =
+      def values[A](model: ConditionalProbabilityTable[A, _]): IndexedSeq[A] =
         model.values
 
-      def combine[A, V](modelsToProbabilities: Map[ConditionalProbabilityTable0[A, V], V])(implicit fieldV: Field[V]): ConditionalProbabilityTable0[A, V] = {
+      def combine[A, V](modelsToProbabilities: Map[ConditionalProbabilityTable[A, V], V])(implicit fieldV: Field[V]): ConditionalProbabilityTable[A, V] = {
 
         val parts: IndexedSeq[(A, V)] =
           modelsToProbabilities.toVector flatMap {
@@ -44,34 +44,34 @@ object ConditionalProbabilityTable0 {
 
         val v = modelsToProbabilities.headOption.map(_._1.variable).getOrElse(Variable[A]("?"))
 
-        ConditionalProbabilityTable0[A, V](newDist, v)
+        ConditionalProbabilityTable[A, V](newDist, v)
       }
 
-      def condition[A, V](model: ConditionalProbabilityTable0[A, V], given: A): ConditionalProbabilityTable0[A, V] =
+      def condition[A, V](model: ConditionalProbabilityTable[A, V], given: A): ConditionalProbabilityTable[A, V] =
         model // TODO true unless G =:= A and model.variable === variable
 
-      def conditionExpression[A, B, V](model: ConditionalProbabilityTable0[A, V], predicate: A => Boolean, screen: A => B): ConditionalProbabilityTable0[B, V] =
+      def conditionExpression[A, B, V](model: ConditionalProbabilityTable[A, V], predicate: A => Boolean, screen: A => B): ConditionalProbabilityTable[B, V] =
         ???
 
-      def empty[A, V](variable: Variable[A])(implicit ringV: Ring[V]): ConditionalProbabilityTable0[A, V] =
-        ConditionalProbabilityTable0(Map.empty, variable)
+      def empty[A, V](variable: Variable[A])(implicit ringV: Ring[V]): ConditionalProbabilityTable[A, V] =
+        ConditionalProbabilityTable(Map.empty, variable)
 
-      def observe[A, V](model: ConditionalProbabilityTable0[A, V], gen: Generator)(implicit spireDist: Dist[V], ringV: Ring[V], orderV: Order[V]): A = {
+      def observe[A, V](model: ConditionalProbabilityTable[A, V], gen: Generator)(implicit spireDist: Dist[V], ringV: Ring[V], orderV: Order[V]): A = {
         val r: V = gen.next[V]
         model.bars.find({ case (_, v) => orderV.gteqv(v, r) }).get._1 // otherwise malformed distribution
       }
 
-      def probabilityOf[A, V](model: ConditionalProbabilityTable0[A, V], a: A)(implicit fieldV: Field[V]): V =
+      def probabilityOf[A, V](model: ConditionalProbabilityTable[A, V], a: A)(implicit fieldV: Field[V]): V =
         model.p.get(a).getOrElse(fieldV.zero)
 
-      def probabilityOfExpression[A, V](model: ConditionalProbabilityTable0[A, V], predicate: A => Boolean)(implicit fieldV: Field[V]): V =
+      def probabilityOfExpression[A, V](model: ConditionalProbabilityTable[A, V], predicate: A => Boolean)(implicit fieldV: Field[V]): V =
         ???
   }
 
 
 }
 
-case class ConditionalProbabilityTable0[A, V](
+case class ConditionalProbabilityTable[A, V](
   p:        Map[A, V],
   variable: Variable[A])(implicit ringV: Ring[V]) {
 
@@ -81,15 +81,5 @@ case class ConditionalProbabilityTable0[A, V](
 
   def charWidth(implicit sa: Show[A]): Int =
     (values.map(a => Show[A].show(a).length).toList).reduce(math.max)
-
-}
-
-case class ConditionalProbabilityTable2[A, G1, G2, N: Field: Order](
-  p:        Map[(G1, G2), Map[A, N]],
-  variable: Variable[A]) {
-
-  lazy val _values = p.values.map(_.keySet).reduce(_ union _).toVector
-
-  def values: IndexedSeq[A] = _values
 
 }