+++ axle-core/src/test/scala/axle/stats/Stats101.scala
@@ -13,7 +13,7 @@ class Stats101 extends FunSuite with Matchers {
 
     val model = uniformDistribution(List(2d, 4d, 4d, 4d, 5d, 5d, 7d, 9d), Variable[Double]("x"))
 
-    standardDeviation[ConditionalProbabilityTable0, Double, Rational](model) should be(2d)
+    standardDeviation[ConditionalProbabilityTable, Double, Rational](model) should be(2d)
   }
 
   test("standard deviation on a list of reals") {
@@ -28,7 +28,7 @@ class Stats101 extends FunSuite with Matchers {
       Real(7),
       Real(9)), Variable[Real]("x"))
 
-    standardDeviation[ConditionalProbabilityTable0, Real, Rational](dist) should be(Real(2))
+    standardDeviation[ConditionalProbabilityTable, Real, Rational](dist) should be(Real(2))
   }
 
 }