+++ axle-wheel/src/main/scala/axle/game/prisoner/package.scala
@@ -3,7 +3,7 @@ package axle.game
 import cats.implicits._
 import spire.math.Rational
 import spire.random.Dist
-import axle.stats.ConditionalProbabilityTable0
+import axle.stats.ConditionalProbabilityTable
 import axle.stats.rationalProbabilityDist
 
 /**
@@ -31,7 +31,7 @@ package object prisoner {
 
       def strategyFor(
         game:   PrisonersDilemma,
-        player: Player): (PrisonersDilemma, PrisonersDilemmaState) => ConditionalProbabilityTable0[PrisonersDilemmaMove, Rational] =
+        player: Player): (PrisonersDilemma, PrisonersDilemmaState) => ConditionalProbabilityTable[PrisonersDilemmaMove, Rational] =
         player match {
           case game.p1 => game.p1Strategy
           case game.p2 => game.p2Strategy