+++ axle-docs/src/main/mdoc/ReleaseNotes.md
@@ -56,7 +56,7 @@ See [Road Map](/road_map/) for the plan of upcoming releases and features.
 ## 0.3.0 (April 12, 2017)
 
 * Scala org to Typelevel
-* Fix malformed distribution in ConditionalProbabilityTable0 and TallyDistribution0
+* Fix malformed distribution in ConditionalProbabilityTable and TallyDistribution0
 * Depend on Spire 0.14.1 (fix mistaken dependency on snapshot release in 0.2.8)
 
 ## 0.2.8 (March 28, 2016)